+++ changelog.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### enhancements
 - wrap webview console as soon as posible to avoid missing errors on logs.
+- wait some seconds before remove media players to avoid shifting on chrome.
 
 ### fix
 - panic on media module when loading initial devices.